backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-docx==0.8.11
pymongo==4.6.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0 